+++ ShoppingCart/Repository/IOrderRepository.cs
@@ -6,12 +6,12 @@ namespace ShoppingCart.Repository
 {
     public interface IOrderRepository
     {
-        Task<Order> GetOrderByIdAsync(int id);
+        Task<Order> GetOrderByIdAsync(long id);
         Task<List<Order>> GetAllOrdersAsync();
         Task<List<Order>> GetAllOrdersByUserIdAsync(string userId);
         Task<Order> AddOrderAsync(Order order);
         Task<Order> UpdateOrderAsync(Order order);
         Task DeleteOrderAsync(Order order);
-        Task DeleteOrderByIdAsync(int id);
+        Task DeleteOrderByIdAsync(long id);
     }
 }
