+++ ShoppingCart/Repository/IProductRepository.cs
@@ -7,11 +7,11 @@ namespace ShoppingCart.Repository
     public interface IProductRepository
     {
         Task<List<Product>> GetAllProductsAsync();
-        Task<Product> GetProductByIdAsync(int productId);
+        Task<Product> GetProductByIdAsync(long productId);
         Task<Product> AddProductAsync(Product product);
         Task<Product> UpdateProductAsync(Product product);
-        Task DeleteProductAsync(int productId);
+        Task DeleteProductAsync(long productId);
         Task<List<Product>> GetProductsByCategorySlugAsync(string slug);
-        Task<List<Product>> GetProductsByCategoryIdAsync(int categoryId);
+        Task<List<Product>> GetProductsByCategoryIdAsync(long categoryId);
     }
 }
