+++ ShoppingCart/Services/IOrderService.cs
@@ -6,9 +6,9 @@ namespace ShoppingCart.Services
     public interface IOrderService
     {
         Task<List<Order>> GetAllOrdersAsync();
-        Task<Order> GetOrderByIdAsync(int id);
+        Task<Order> GetOrderByIdAsync(long id);
         Task<Order> CreateOrderAsync(Order order);
         Task<Order> UpdateOrderAsync(Order order);
-        Task DeleteOrderAsync(int id);
+        Task DeleteOrderAsync(long id);
     }
 }
