+++ ShoppingCart/Services/IProductService.cs
@@ -7,13 +7,13 @@ namespace ShoppingCart.Services
     public interface IProductService
     {
         Task<List<Product>> GetAllProductsAsync();
-        Task<Product> GetProductByIdAsync(int productId);
+        Task<Product> GetProductByIdAsync(long productId);
         Task<Product> CreateProductAsync(Product product);
         Task<Product> UpdateProductAsync(Product product);
-        Task DeleteProductAsync(int productId);
+        Task DeleteProductAsync(long productId);
         Task<List<Product>> GetProductsByCategorySlugAsync(string categorySlug);
         Task<Category> GetCategoryBySlugAsync(string slug);
-        Task<List<Product>> GetProductsByCategoryIdAsync(int categoryId);
+        Task<List<Product>> GetProductsByCategoryIdAsync(long categoryId);
 
     }
 }
